refactor(types): extract source and notebook models
